+++ src/extension/V425fixable.ts
@@ -7,6 +7,7 @@ import { fixWrongCppStandard } from "./fixes/wrongCppStandard";
 
 import * as console from "../console";
 import { fixWrongIntellisenseMode } from "./fixes/wrongIntellisenseMode";
+import { fixPropCompilerPath } from "./fixes/propCompilerPath";
 
 
 export class V425Fixable extends Fixable {
@@ -36,6 +37,8 @@ export class V425Fixable extends Fixable {
 
         fixWrongIntellisenseMode(this.project);
         console.log("End fix wrong intellisense mode.\n");
+        
+        fixPropCompilerPath(this.project);
         
         return;
     }